requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
apify-client==1.7.0
apify-shared==1.1.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
exceptiongroup==1.2.1
FLAML==2.1.2
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-resources==6.4.0
jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
markdownify==0.12.1
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.4
openai==1.30.1
packaging==24.0
pandas==2.0.3
pillow==10.3.0
pkgutil-resolve-name==1.3.10
protobuf==4.25.3
pyarrow==16.1.0
pyautogen==0.2.27
pydantic==2.7.1
pydantic-core==2.18.2
pydeck==0.9.1
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
streamlit==1.34.0
tenacity==8.3.0
termcolor==2.4.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
zipp==3.18.2
